use size_t to avoid compiler warning
Former-commit-id: 8c1b05d0baba9299efa9df5b39ed05b8e130f3d4

+++ C/hel.cpp
@@ -8,7 +8,7 @@ int main(void)
 	string foo = "abcdef";
 	char ch;
 
-	for (int i = 0; i < foo.length(); i++)
+	for (size_t i = 0; i < foo.length(); i++)
 	{
 		string bar = foo.substr(i, 1);
 		ch = bar.c_str()[0];
